+++ src/store/actionsType/actionsType.js
@@ -1,0 +1,4 @@
+export const START_CHART = "START_CHART";
+export const STOP_CHART = "STOP_CHART";
+export const ADD_COIN = "ADD_COIN";
+export const DELETE_COIN = "DELETE_COIN";
