+++ src/store/store.js
@@ -1,6 +1,5 @@
 import { applyMiddleware, combineReducers, createStore, compose } from "redux";
 import thunk from "redux-thunk";
-import { coinReducer } from "./reducers/coin-reducer";
 import { chartReducer } from "./reducers/chart-reducer";
 
 const reducers = combineReducers({ chart: chartReducer });
@@ -10,4 +9,3 @@ export const store = createStore(reducers,
         applyMiddleware(thunk),
         window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()));
 
-window.store = store;
